[ocean] Sync repos
Summary:
Syncing repos.

fbshipit-source-id: ce1dc406a9a5b3bf7540daabd7a6d3725cf52f7f

+++ impl/ocean/network/ConnectionOrientedClient.cpp
@@ -123,7 +123,7 @@ ConnectionOrientedClient::SocketResult ConnectionOrientedClient::send(const void
 		}
 	}
 #else
-	if (result == -1)
+	if (result == size_t(-1))
 	{
 		if (errno == EWOULDBLOCK)
 		{
